requirements.txt
python-dotenv
langchain-nvidia-ai-endpoints
PyPDF2

# FastAPI Stack
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
aiofiles>=23.2.1

# NeMo Guardrails
nemoguardrails>=0.10.0

# PDF Redaction
pymupdf>=1.23.0

# RAG (Retrieval-Augmented Generation) Stack
langchain>=0.1.0
chromadb>=0.4.0  # Vector database for semantic search
sentence-transformers>=2.2.0  # Local embeddings (sovereign/offline mode)
tiktoken>=0.5.0  # Token counting for chunking

# Nemotron Parse (document parsing with bounding boxes)
pdf2image>=1.16.0  # Convert PDF pages to images for Nemotron
httpx>=0.25.0  # Async HTTP client for API calls
Pillow>=10.0.0  # Image processing